+++ lib/src/circle_percentage.dart
@@ -28,7 +28,7 @@ class CirclePercentage extends StatefulWidget {
   final String? centerText;
 
   /// This parameter is the text style of the label for the percentage text
-  final TextStyle textStyle;
+  final TextStyle centerTextStyle;
 
   const CirclePercentage({
     super.key,
@@ -40,7 +40,7 @@ class CirclePercentage extends StatefulWidget {
     this.percentageColor = Colors.black,
     this.backgroundColor = Colors.black12,
     this.centerText,
-    this.textStyle = const TextStyle(color: Colors.black),
+    this.centerTextStyle = const TextStyle(color: Colors.black),
   })  : assert(currentPercentage <= maxPercentage),
         assert(currentPercentage >= 0),
         assert(duration >= 0);
@@ -81,7 +81,7 @@ class _CirclePercentageState extends State<CirclePercentage> {
                 widget.centerText ??
                     (value * widget.maxPercentage).toInt().toString(),
                 textAlign: TextAlign.center,
-                style: widget.textStyle,
+                style: widget.centerTextStyle,
               ),
             ),
           ),